chore(testing): only compile ValidatableComponent in test builds
