+++ src/sinks/splunk_hec/logs/config.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use vector_lib::{
-    codecs::{JsonSerializerConfig, MetricTagValues, TextSerializerConfig},
+    codecs::TextSerializerConfig,
     lookup::lookup_v2::{ConfigValuePath, OptionalTargetPath},
     sensitive_string::SensitiveString,
 };
@@ -280,73 +280,75 @@ impl HecLogsSinkConfig {
     }
 }
 
-impl ValidatableComponent for HecLogsSinkConfig {
-    fn validation_configuration() -> ValidationConfiguration {
-        let endpoint = "http://127.0.0.1:9001".to_string();
-
-        let mut batch = BatchConfig::default();
-        batch.max_events = Some(1);
-
-        let config = Self {
-            endpoint: endpoint.clone(),
-            default_token: "i_am_an_island".to_string().into(),
-            host_key: config_host_key_target_path(),
-            indexed_fields: vec![],
-            index: None,
-            sourcetype: None,
-            source: None,
-            encoding: EncodingConfig::new(
-                JsonSerializerConfig::new(MetricTagValues::Full).into(),
-                Transformer::default(),
-            ),
-            compression: Compression::default(),
-            batch,
-            request: TowerRequestConfig {
-                timeout_secs: 2,
-                retry_attempts: 0,
-                ..Default::default()
-            },
-            tls: None,
-            acknowledgements: HecClientAcknowledgementsConfig {
-                indexer_acknowledgements_enabled: false,
-                ..Default::default()
-            },
-            timestamp_nanos_key: None,
-            timestamp_key: config_timestamp_key_target_path(),
-            auto_extract_timestamp: None,
-            endpoint_target: EndpointTarget::Raw,
-        };
-
-        let endpoint = format!("{endpoint}/services/collector/raw");
-
-        let external_resource = ExternalResource::new(
-            ResourceDirection::Push,
-            HttpResourceConfig::from_parts(
-                http::Uri::try_from(&endpoint).expect("should not fail to parse URI"),
-                None,
-            ),
-            config.encoding.clone(),
-        );
-
-        ValidationConfiguration::from_sink(
-            Self::NAME,
-            vec![ComponentTestCaseConfig::from_sink(
-                config,
-                None,
-                Some(external_resource),
-            )],
-        )
-    }
-}
-
-register_validatable_component!(HecLogsSinkConfig);
-
 #[cfg(test)]
 mod tests {
-    use super::HecLogsSinkConfig;
+    use super::*;
+    use crate::components::validation::prelude::*;
+    use vector_lib::codecs::{JsonSerializerConfig, MetricTagValues};
 
     #[test]
     fn generate_config() {
         crate::test_util::test_generate_config::<HecLogsSinkConfig>();
     }
+
+    impl ValidatableComponent for HecLogsSinkConfig {
+        fn validation_configuration() -> ValidationConfiguration {
+            let endpoint = "http://127.0.0.1:9001".to_string();
+
+            let mut batch = BatchConfig::default();
+            batch.max_events = Some(1);
+
+            let config = Self {
+                endpoint: endpoint.clone(),
+                default_token: "i_am_an_island".to_string().into(),
+                host_key: config_host_key_target_path(),
+                indexed_fields: vec![],
+                index: None,
+                sourcetype: None,
+                source: None,
+                encoding: EncodingConfig::new(
+                    JsonSerializerConfig::new(MetricTagValues::Full).into(),
+                    Transformer::default(),
+                ),
+                compression: Compression::default(),
+                batch,
+                request: TowerRequestConfig {
+                    timeout_secs: 2,
+                    retry_attempts: 0,
+                    ..Default::default()
+                },
+                tls: None,
+                acknowledgements: HecClientAcknowledgementsConfig {
+                    indexer_acknowledgements_enabled: false,
+                    ..Default::default()
+                },
+                timestamp_nanos_key: None,
+                timestamp_key: config_timestamp_key_target_path(),
+                auto_extract_timestamp: None,
+                endpoint_target: EndpointTarget::Raw,
+            };
+
+            let endpoint = format!("{endpoint}/services/collector/raw");
+
+            let external_resource = ExternalResource::new(
+                ResourceDirection::Push,
+                HttpResourceConfig::from_parts(
+                    http::Uri::try_from(&endpoint).expect("should not fail to parse URI"),
+                    None,
+                ),
+                config.encoding.clone(),
+            );
+
+            ValidationConfiguration::from_sink(
+                Self::NAME,
+                vec![ComponentTestCaseConfig::from_sink(
+                    config,
+                    None,
+                    Some(external_resource),
+                )],
+            )
+        }
+    }
+
+    register_validatable_component!(HecLogsSinkConfig);
 }
