+++ src/sources/http_client/client.rs
@@ -10,6 +10,7 @@ use snafu::ResultExt;
 use std::{collections::HashMap, time::Duration};
 use tokio_util::codec::Decoder as _;
 
+use crate::sources::util::http_client;
 use crate::{
     codecs::{Decoder, DecodingConfig},
     config::{SourceConfig, SourceContext},
@@ -26,7 +27,6 @@ use crate::{
     tls::{TlsConfig, TlsSettings},
     Result,
 };
-use crate::{components::validation::prelude::*, sources::util::http_client};
 use vector_lib::codecs::{
     decoding::{DeserializerConfig, FramingConfig},
     StreamDecodingError,
@@ -234,37 +234,6 @@ impl SourceConfig for HttpClientConfig {
         false
     }
 }
-
-impl ValidatableComponent for HttpClientConfig {
-    fn validation_configuration() -> ValidationConfiguration {
-        let uri = Uri::from_static("http://127.0.0.1:9898");
-
-        let config = Self {
-            endpoint: uri.to_string(),
-            interval: Duration::from_secs(1),
-            timeout: Duration::from_secs(1),
-            decoding: DeserializerConfig::Json(Default::default()),
-            ..Default::default()
-        };
-
-        let external_resource = ExternalResource::new(
-            ResourceDirection::Pull,
-            HttpResourceConfig::from_parts(uri, Some(config.method.into())),
-            config.get_decoding_config(None),
-        );
-
-        ValidationConfiguration::from_source(
-            Self::NAME,
-            vec![ComponentTestCaseConfig::from_source(
-                config,
-                None,
-                Some(external_resource),
-            )],
-        )
-    }
-}
-
-register_validatable_component!(HttpClientConfig);
 
 impl HttpClientConfig {
     pub fn get_decoding_config(&self, log_namespace: Option<LogNamespace>) -> DecodingConfig {
